# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Family Locator API — Life360-like app
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /circles:
    get:
      operationId: listCircles
      tags: [circles]
      summary: List circles the current user belongs to
      responses:
        "200":
          description: List of circles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Circle"
    post:
      operationId: createCircle
      tags: [circles]
      summary: Create a new circle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CircleInput"
      responses:
        "201":
          description: Created circle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Circle"
  /circles/{circleId}:
    get:
      operationId: getCircle
      tags: [circles]
      summary: Get a circle by ID
      parameters:
        - name: circleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Circle details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Circle"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateCircle
      tags: [circles]
      summary: Update a circle
      parameters:
        - name: circleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CircleUpdate"
      responses:
        "200":
          description: Updated circle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Circle"
    delete:
      operationId: deleteCircle
      tags: [circles]
      summary: Delete a circle
      parameters:
        - name: circleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /circles/{circleId}/summary:
    get:
      operationId: getCircleSummary
      tags: [circles]
      summary: Get aggregated stats for a circle (member count, place count, recent events)
      parameters:
        - name: circleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Circle summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CircleSummary"

  # ── Members ──────────────────────────────────────────────────────────────
  /circles/{circleId}/members:
    get:
      operationId: listCircleMembers
      tags: [members]
      summary: List members of a circle with their latest location
      parameters:
        - name: circleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
  /circles/{circleId}/members/invite:
    post:
      operationId: inviteMember
      tags: [members]
      summary: Invite a user to a circle by email
      parameters:
        - name: circleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteInput"
      responses:
        "201":
          description: Invitation sent / member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "404":
          description: Circle or user not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /circles/{circleId}/members/{memberId}:
    delete:
      operationId: removeCircleMember
      tags: [members]
      summary: Remove a member from a circle
      parameters:
        - name: circleId
          in: path
          required: true
          schema:
            type: integer
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # ── Locations ─────────────────────────────────────────────────────────────
...
  /circles/{circleId}/places:
    get:
      operationId: listPlaces
      tags: [places]
      summary: List saved places for a circle
      parameters:
        - name: circleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of places
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Place"
    post:
      operationId: createPlace
      tags: [places]
      summary: Create a saved place for a circle
      parameters:
        - name: circleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceInput"
      responses:
        "201":
          description: Created place
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Place"
  /circles/{circleId}/places/{placeId}:
    patch:
      operationId: updatePlace
      tags: [places]
      summary: Update a saved place
      parameters:
        - name: circleId
          in: path
          required: true
          schema:
            type: integer
        - name: placeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceUpdate"
      responses:
        "200":
          description: Updated place
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Place"
    delete:
      operationId: deletePlace
      tags: [places]
      summary: Delete a saved place
      parameters:
        - name: circleId
          in: path
          required: true
          schema:
            type: integer
        - name: placeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Activity ──────────────────────────────────────────────────────────────
  /circles/{circleId}/activity:
    get:
      operationId: getCircleActivity
      tags: [activity]
      summary: Get recent activity events for a circle (arrivals, departures, check-ins)
      parameters:
        - name: circleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Activity events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityEvent"
  /circles/{circleId}/checkin:
    post:
      operationId: checkIn
      tags: [activity]
      summary: Post a check-in for a member at a location
      parameters:
        - name: circleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckInInput"
      responses:
        "201":
          description: Check-in recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityEvent"
...
  /circles/{circleId}/messages:
    get:
      operationId: listCircleMessages
      tags: [circles]
      summary: Get messages for a circle
      parameters:
        - name: circleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CircleMessage"
    post:
      operationId: sendCircleMessage
      tags: [circles]
      summary: Send a message to a circle
      parameters:
        - name: circleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CircleMessage"
  /circles/{circleId}/members/{memberId}/trail:
    get:
      operationId: getMemberTrail
      tags: [members]
      summary: Get recent location trail for a circle member (last 10 points)
      parameters:
        - name: circleId
          in: path
          required: true
          schema:
            type: integer
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Location trail
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LocationRecord"
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Circles ───────────────────────────────────────────────────────────────
    Circle:
      type: object
      required: [id, name, color, ownerId, memberCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: [string, 'null']
        color:
          type: string
        ownerId:
          type: string
        memberCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    CircleInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        color:
          type: string
    CircleUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        color:
          type: string
    CircleSummary:
      type: object
      required: [circleId, memberCount, placeCount, recentEventCount, onlineCount]
      properties:
        circleId:
          type: integer
        memberCount:
          type: integer
        placeCount:
          type: integer
        recentEventCount:
          type: integer
        onlineCount:
          type: integer

    # ── Members ───────────────────────────────────────────────────────────────
    Member:
      type: object
      required: [id, userId, circleId, role, joinedAt, user]
      properties:
        id:
          type: integer
        userId:
          type: string
        circleId:
          type: integer
        role:
          type: string
          enum: [owner, member]
        joinedAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/AuthUser"
        lastLocation:
          oneOf:
            - $ref: "#/components/schemas/LocationRecord"
            - type: 'null'
    InviteInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email

    # ── Locations ─────────────────────────────────────────────────────────────
...
    Place:
      type: object
      required: [id, circleId, name, latitude, longitude, radius, icon, createdAt]
      properties:
        id:
          type: integer
        circleId:
          type: integer
        name:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        radius:
          type: number
        icon:
          type: string
        createdAt:
          type: string
          format: date-time
    PlaceInput:
      type: object
      required: [name, latitude, longitude]
      properties:
        name:
          type: string
          minLength: 1
        latitude:
          type: number
        longitude:
          type: number
        radius:
          type: number
        icon:
          type: string
    PlaceUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        latitude:
          type: number
        longitude:
          type: number
        radius:
          type: number
        icon:
          type: string

    # ── SOS ───────────────────────────────────────────────────────────────────
...
    CheckInInput:
      type: object
      required: [description]
      properties:
        description:
          type: string
          minLength: 1
        placeId:
          type: integer

    # ── Chat ─────────────────────────────────────────────────────────────────
...